requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
pytest
pytest-django
requests
openai
python-dotenv
